Refactor board logic to use integer coordinates throughout
Refactored Board, Game, and Gui classes to use sf::Vector2i for all board positions, replacing sf::Vector2f. Updated move validation and drawing logic to use integer coordinates, improving type safety and consistency. Changed getPossibleSquares and getAttackedSquares to use integer types. Game::getBoard now returns a pointer, and a changePlayerColor method was added. GUI logic now handles board orientation for both players and includes additional debug output. These changes lay the groundwork for more robust move validation and interaction.

+++ Chess/board.h
@@ -9,8 +9,8 @@ public:
     bool move(int fromX, int fromY, int toX, int toY);
     Piece getPiece(int x, int y) const;
     bool isKingChecked(PieceColor pieceColor) const;
-    std::vector<sf::Vector2f> getPossibleSquares(sf::Vector2f selectedSquare) const;
-    std::vector<sf::Vector2f> getAttackedSquares(sf::Vector2f selectedSquare) const;
+    std::vector<sf::Vector2i> getPossibleSquares(sf::Vector2i selectedSquare) const;
+    std::vector<sf::Vector2i> getAttackedSquares(sf::Vector2i selectedSquare) const;
 
 private:
     Piece board[8][8];
